+++ src/action/config/ActionConfiguration.cpp
@@ -25,9 +25,9 @@ ActionConfiguration::ActionConfiguration(
     const mesh::PtrMeshConfiguration &meshConfig)
     : NAME_DIVIDE_BY_AREA("divide-by-area"),
       NAME_MULTIPLY_BY_AREA("multiply-by-area"),
-      NAME_SCALE_BY_COMPUTED_DT_RATIO("scale-by-computed-dt-ratio"),
-      NAME_SCALE_BY_COMPUTED_DT_PART_RATIO("scale-by-computed-dt-part-ratio"),
-      NAME_SCALE_BY_DT("scale-by-dt"),
+      NAME_SCALING_BY_TIME_STEP_TO_TIME_WINDOW_RATIO("scale-by-computed-dt-ratio"),       //@todo rename this, breaking change!
+      NAME_SCALING_BY_COMPUTED_TIME_WINDOW_PART_RATIO("scale-by-computed-dt-part-ratio"), //@todo rename this, breaking change!
+      NAME_SCALING_BY_TIME_WINDOW_SIZE("scale-by-dt"),                                    //@todo rename this, breaking change!, currently misleading. See https://github.com/precice/precice/issues/934
       NAME_SUMMATION("summation"),
       NAME_COMPUTE_CURVATURE("compute-curvature"),
       NAME_PYTHON("python"),
@@ -77,24 +77,24 @@ ActionConfiguration::ActionConfiguration(
     tags.push_back(tag);
   }
   {
-    XMLTag tag(*this, NAME_SCALE_BY_COMPUTED_DT_RATIO, occ, TAG);
-    tag.setDocumentation("Multiplies source data values by ratio of full dt / last computed dt,"
+    XMLTag tag(*this, NAME_SCALING_BY_TIME_STEP_TO_TIME_WINDOW_RATIO, occ, TAG);
+    tag.setDocumentation("Multiplies source data values by ratio of last time step size / time window size,"
                          " and writes the result into target data.");
     tag.addSubtag(tagSourceData);
     tag.addSubtag(tagTargetData);
     tags.push_back(tag);
   }
   {
-    XMLTag tag(*this, NAME_SCALE_BY_COMPUTED_DT_PART_RATIO, occ, TAG);
-    tag.setDocumentation("Multiplies source data values by ratio of full dt / computed dt part,"
+    XMLTag tag(*this, NAME_SCALING_BY_COMPUTED_TIME_WINDOW_PART_RATIO, occ, TAG);
+    tag.setDocumentation("Multiplies source data values by ratio of computed time window part / time window size,"
                          " and writes the result into target data.");
     tag.addSubtag(tagSourceData);
     tag.addSubtag(tagTargetData);
     tags.push_back(tag);
   }
   {
-    XMLTag tag(*this, NAME_SCALE_BY_DT, occ, TAG);
-    tag.setDocumentation("Multiplies source data values by last computed dt, and writes the "
+    XMLTag tag(*this, NAME_SCALING_BY_TIME_WINDOW_SIZE, occ, TAG);
+    tag.setDocumentation("Multiplies source data values by the time window size, and writes the "
                          "result into target data.");
     tag.addSubtag(tagSourceData);
     tag.addSubtag(tagTargetData);
@@ -259,18 +259,18 @@ void ActionConfiguration::createAction()
     action = action::PtrAction(
         new action::ScaleByAreaAction(timing, targetDataID,
                                       mesh, action::ScaleByAreaAction::SCALING_DIVIDE_BY_AREA));
-  } else if (_configuredAction.type == NAME_SCALE_BY_COMPUTED_DT_RATIO) {
+  } else if (_configuredAction.type == NAME_SCALING_BY_TIME_STEP_TO_TIME_WINDOW_RATIO) {
     action = action::PtrAction(
         new action::ScaleByDtAction(timing, sourceDataIDs.back(), targetDataID,
-                                    mesh, action::ScaleByDtAction::SCALING_BY_COMPUTED_DT_RATIO));
-  } else if (_configuredAction.type == NAME_SCALE_BY_COMPUTED_DT_PART_RATIO) {
+                                    mesh, action::ScaleByDtAction::SCALING_BY_TIME_STEP_TO_TIME_WINDOW_RATIO));
+  } else if (_configuredAction.type == NAME_SCALING_BY_COMPUTED_TIME_WINDOW_PART_RATIO) {
     action = action::PtrAction(
         new action::ScaleByDtAction(timing, sourceDataIDs.back(), targetDataID,
-                                    mesh, action::ScaleByDtAction::SCALING_BY_COMPUTED_DT_PART_RATIO));
-  } else if (_configuredAction.type == NAME_SCALE_BY_DT) {
+                                    mesh, action::ScaleByDtAction::SCALING_BY_COMPUTED_TIME_WINDOW_PART_RATIO));
+  } else if (_configuredAction.type == NAME_SCALING_BY_TIME_WINDOW_SIZE) {
     action = action::PtrAction(
         new action::ScaleByDtAction(timing, sourceDataIDs.back(), targetDataID,
-                                    mesh, action::ScaleByDtAction::SCALING_BY_DT));
+                                    mesh, action::ScaleByDtAction::SCALING_BY_TIME_WINDOW_SIZE));
   } else if (_configuredAction.type == NAME_COMPUTE_CURVATURE) {
     action = action::PtrAction(
         new action::ComputeCurvatureAction(timing, targetDataID,