+++ src/action/tests/ScaleActionTest.cpp
@@ -62,7 +62,7 @@ BOOST_AUTO_TEST_CASE(DivideByArea)
   BOOST_TEST(values(2) == 8.0);
 }
 
-BOOST_AUTO_TEST_CASE(ScaleByComputedTimestepLength)
+BOOST_AUTO_TEST_CASE(ScaleByTimeStepSizeToTimeWindowSize)
 {
   PRECICE_TEST(1_rank);
   using namespace mesh;
@@ -83,7 +83,7 @@ BOOST_AUTO_TEST_CASE(ScaleByComputedTimestepLength)
 
   action::ScaleByDtAction scale(
       action::ScaleByDtAction::WRITE_MAPPING_POST, sourceDataID, targetDataID, mesh,
-      action::ScaleByDtAction::SCALING_BY_COMPUTED_DT_RATIO);
+      action::ScaleByDtAction::SCALING_BY_TIME_STEP_TO_TIME_WINDOW_RATIO);
 
   scale.performAction(0.0, 0.0, 0.0, 1.0);
   BOOST_TEST(sourceValues(0) == 2.0);
@@ -118,7 +118,7 @@ BOOST_AUTO_TEST_CASE(ScaleByComputedTimestepLength)
   BOOST_TEST(targetValues(2) == 1.0);
 }
 
-BOOST_AUTO_TEST_CASE(ScaleByComputedTimestepPartLength)
+BOOST_AUTO_TEST_CASE(ScaleByComputedTimeWindowPart)
 {
   PRECICE_TEST(1_rank);
   using namespace mesh;
@@ -138,7 +138,7 @@ BOOST_AUTO_TEST_CASE(ScaleByComputedTimestepPartLength)
 
   action::ScaleByDtAction scale(
       action::ScaleByDtAction::WRITE_MAPPING_POST, sourceDataID, targetDataID, mesh,
-      action::ScaleByDtAction::SCALING_BY_COMPUTED_DT_PART_RATIO);
+      action::ScaleByDtAction::SCALING_BY_COMPUTED_TIME_WINDOW_PART_RATIO);
 
   scale.performAction(0.0, 0.0, 0.0, 1.0);
   BOOST_TEST(sourceValues(0) == 2.0);